+++ src/forms/CommentForm/CommentForm.scss
@@ -1,2 +1,17 @@
-.commentform {
+.newcomment {
+  display: grid;
+  align-items: flex-start;
+  padding-bottom: 0.25rem;
+  grid-template-columns: min-content auto min-content;
+  column-gap: 0.75rem;
+  .input {
+    flex-grow: 1;
+  }
+  .avatar {
+    // margin-right: 0.5rem;
+    margin-top: 0.45rem;
+  }
+  .button {
+    margin-top: 0.4rem;
+  }
 }
