--- src/templates/NewComment/NewComment.md
@@ -1,10 +1,0 @@
-#### Basic Usage
-
-```jsx
-<NewComment>NewComment</NewComment>
-```
-
-Source:
-
-```js { "file": "./NewComment.js" }
-```
