--- src/templates/NewComment/NewComment.scss
@@ -1,15 +1,0 @@
-.newcomment {
-  display: flex;
-  align-items: flex-start;
-  padding-bottom: 0.25rem;
-  .input {
-    flex-grow: 1;
-  }
-  .avatar {
-    margin-right: 0.5rem;
-    margin-top: 0.3rem;
-  }
-  .button {
-    margin-top: 0.2rem;
-  }
-}
